feat(kernel module): rewrite kernel module to use a char device instead of kprobes
The kernel module uses a virtual char device to apply the session semantics. At this point the
device can be registered and writing or reading from the device filename will make the userspac
application interact with the path that enables the session semantics.

BREAKING CHANGE: removed kernel modules in favor of a char device

re #1

Signed-off-by: Nick1296 <mattianicolella@gmail.com>