improvement(char device): remove a race condition on device removal
During the device removal we could have had another process that was operating on the device, now we
have a refcount anda flag, which are used to check if we have processes using the device and to
avoid using a device while is being removed. A device cannot be removed if there are active session.
If there is an error during the removal of the session EPIPE is returned and SIGPIPE is sent (or at
least we try to send it). Finally a bug on deallocation was resolved.

Signed-off-by: Nick1296 <mattianicolella@gmail.com>