improvement(session manager): remove a race condition when removing an element from the rculist
The rcu list now contains objects that have a pointer to the session element, to avoid sideffects
when modifying a session. Session objects now have a refcount that keeps track of the processes that
are using the session. The deallocation of a session object now happens only when the refcount is
<=1 (only the current process, which is attempting the removal, is using the object).

Signed-off-by: Nick1296 <mattianicolella@gmail.com>