fix(char device, libsessionfs, session manager): fix close semantic for non-session files
The semantic of the close can now discern files that use session semantic and file that use normal
semantic. This has been possible since the library now will examine the file path before calling the
module, making possible to determine if the module must be called or not and allowing the module to
search for a session using the file path even when a close is called.

Signed-off-by: Nick1296 <mattianicolella@gmail.com>