fix(single-thread test): complete single thread test, add support for multi process test
The module has been tested with a single process that work on a psudorandom number of different
files (up to 100), where each file has a pseudorandom dimension up to 1MB. The module also
identifies correctly sessions which have been left opened and is capable to close them before
shutdown. However the module is still prone to race conditions during unloads.

Signed-off-by: Nick1296 <mattianicolella@gmail.com>