fix(memory issues): solve memory issues in shared library and userspace program
Some memory issues in the userspace program were solved usin ASAN, which is now preloaded before our
library. The userspace program now takes the max number of files and processes as arguments and the
makefile was modified to do some stressful test in single and multi process mode. Finally the char
device now inserts a "just in case" terminator at the end of the sess_path variable during the
device write.

Signed-off-by: Nick1296 <mattianicolella@gmail.com>