fix(concurrency errors): fix some concuerrency errors and add file references in previous commit
Some errors related to race conditions were fixed in the device drivers and in the session manager.
In the device driver the path check function now grabs correctly the lock on sess_path. In the
session manager many critical section have been reduced and during the file copy preemption is now
disabled, since with large files the scheduler will try to preempt our process while opening or
closing an incarnation, after having grabbed the read-write lock. Additionally the session search
and the session close have been rewritten since they made a wrong use of the locks and the session
refcount. Incidentally now each function that gets a session or creates it checks if the given
session is valid and make a better use of locks.

Signed-off-by: Nick1296 <mattianicolella@gmail.com>