improvement(session manager, module release): prevent module release when in use
Now the module cannot be released anytime, but only when there are no processes using it and no
valid active sessions. To do so the session manager relase has been rewritten and the policy used to
grab locks and increment refcounts has been changed. However the release condition is checked only
after having completed an operation on the device.

Signed-off-by: Nick1296 <mattianicolella@gmail.com>