fix(docs, tests, kmodule): fix rmmod disable, improve docs, fix shared lib and tests
Documentation has been improved and includes a report of the current state of the project. To unload
the module an ioctl is required, which triggers the checks that in the previous versions were made
after avery operation on the device. The shared library now closes the device file after every
operation and some concurrency problems have been addressed in the session manager.

Signed-off-by: Nick1296 <mattianicolella@gmail.com>