fix(kmodule bug, more info in userspace): remove bugs in session_info and reduce critical sections
A use after free bug was removed in the session info submodule, requiring the modification of the
deallocation algorithm for sessions and incarnations. In the session manager critical section now
correctly exclude functions that may sleep, like open_file(). The libsessionfs and userspace program
have now more informative messages and the userspace programs does not leak memory when an assert
fails. Documentation has been updated to match the changes.

Signed-off-by: Nick1296 <mattianicolella@gmail.com>